requirements.txt
flask
waitress
faiss-cpu
numpy
openai
requests
openpyxl
